cleaned up Migrations

+++ lib/database/migrations/2017_02_21_124507_create_county_sub_table.php
@@ -16,7 +16,7 @@ class CreateCountySubTable extends Migration
         Schema::create('tbl_county_sub', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->integer('county_id')->unsigned();
+            $table->bigInteger('county_id')->unsigned();
             $table->timestamps();
             $table->softDeletes();
             $table->foreign('county_id')->references('id')->on('tbl_county');
